+++ src/Table/Dto/Cells/Chip.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 namespace Patrikjak\Utils\Table\Dto\Cells;
 
 use Patrikjak\Utils\Common\Enums\Type;
-use Patrikjak\Utils\Table\Enums\CellType;
+use Patrikjak\Utils\Table\Enums\Cells\CellType;
 use Patrikjak\Utils\Table\Interfaces\Cells\Cell as CellInterface;
 
 class Chip extends Cell implements CellInterface
