+++ src/Table/Dto/Cells/Double.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Dto\Cells;
 
-use Patrikjak\Utils\Table\Enums\CellType;
+use Patrikjak\Utils\Table\Enums\Cells\CellType;
 use Patrikjak\Utils\Table\Interfaces\Cells\Cell as CellInterface;
 
 class Double extends Cell implements CellInterface
