+++ src/Table/Dto/Cells/Simple.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 namespace Patrikjak\Utils\Table\Dto\Cells;
 
 use Patrikjak\Utils\Common\Enums\Icon;
-use Patrikjak\Utils\Table\Enums\CellType;
+use Patrikjak\Utils\Table\Enums\Cells\CellType;
 use Patrikjak\Utils\Table\Interfaces\Cells\Cell as CellInterface;
 use Patrikjak\Utils\Table\Interfaces\Cells\SupportsIcon;
