+++ src/Table/Dto/Filter/Definitions/FilterDefinition.php
@@ -2,7 +2,7 @@
 
 namespace Patrikjak\Utils\Table\Dto\Filter\Definitions;
 
-use Patrikjak\Utils\Table\Dto\Filter\FilterType;
+use Patrikjak\Utils\Table\Enums\Filter\FilterType;
 
 interface FilterDefinition
 {
