+++ src/Table/Dto/Filter/Definitions/Text/TextFilterDefinition.php
@@ -3,7 +3,7 @@
 namespace Patrikjak\Utils\Table\Dto\Filter\Definitions\Text;
 
 use Patrikjak\Utils\Table\Dto\Filter\Definitions\FilterDefinition;
-use Patrikjak\Utils\Table\Dto\Filter\FilterType;
+use Patrikjak\Utils\Table\Enums\Filter\FilterType;
 
 final readonly class TextFilterDefinition implements FilterDefinition
 {
