+++ src/Table/Dto/Sort/SortCriteria.php
@@ -4,6 +4,8 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Dto\Sort;
 
+use Patrikjak\Utils\Table\Enums\Sort\SortOrder;
+
 final readonly class SortCriteria
 {
     public function __construct(
